Fix slow consumer causing deadlock
In the event of a slow consumer (slow http connection) it is possible for
the buffers to get full causing a upstream lock to be held in the counts
code, causing the clustercache event queue to grow very large.  To mitigate
against slow consumers if one is found we stop the subscription now.

Additionally in the situation of a slow consumer, and then the consumer
disconnected, we could get into a deadlock situation during cancelation.
This is because the channel with a full buffer was not being emptied fully.