Give some chance to consumers
Having many events come quickly into the c channel can cause the watch
to be canceled even though the consumer is NOT slow.

To fix this, we give some maximum time for an event to arrive in the
queue, giving a chance to the consumer to consume the event and get
ready again.